added geometry calc file
