+++ README.md
@@ -3,14 +3,14 @@
 ## Description
 This project provides tools for camera calibration and Apriltag pose estimation, tailored for use with the MBot equipped with a **Jetson Nano** and **Raspberry Pi 5**.
 
-In the `jetson_nano_` and `pi5` folders, the scripts offer:
+In the `jetson_nano` and `pi5` folders, the scripts offer:
 - Streaming live camera feeds to a browser for real-time viewing and interaction.
 - Publishing and subscribing to Apriltag LCM messages.
 - MBot following Apriltags.
 
 ## Installation and Usage
 
-Refer to the README.md file in each folder for detailed instructions.
+For detailed setup instructions, refer to the `README.md` in each folder `jetson_nano` or `pi5`. These are standalone projects, so you can remove either one if you don’t need it for your setup.
 
 ## Authors and maintainers
 - The current maintainer of this project is Shaw Sun. Please direct all questions regarding support, contributions, and issues to the maintainer.
